pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "agente-ia"
version = "0.1.0"
description = "Multi-tenant conversational AI agent platform for businesses"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Alexander Mejia", email = "alexander777800@gmail.com"}
]
keywords = ["ai", "saas", "agents", "conversational", "whatsapp", "multi-tenant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "supabase==2.3.5",
    "openai==1.3.0",
    "redis==5.0.1",
    "requests==2.31.0",
    "httpx==0.25.2",
    "python-jose==3.3.0",
    "passlib==1.7.4",
    "cryptography==41.0.7",
    "pyjwt==2.8.1",
    "structlog==23.2.0",
    "python-json-logger==2.0.7",
    "python-dateutil==2.8.2",
    "phonenumbers==8.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-watch==4.2.0",
    "pytest-timeout==2.1.0",
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "isort==5.13.2",
    "bandit==1.7.5",
    "pip-audit==2.6.1",
]

dashboard = [
    "streamlit==1.29.0",
    "plotly==5.18.0",
    "pandas==2.1.4",
]

all = [
    "agente-ia[dev,dashboard]",
]

[project.urls]
Homepage = "https://github.com/yourusername/agente-ia"
Documentation = "https://github.com/yourusername/agente-ia/blob/main/CLAUDE.md"
Repository = "https://github.com/yourusername/agente-ia"
Issues = "https://github.com/yourusername/agente-ia/issues"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [".venv/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "supabase.*",
    "fastapi.*",
    "pydantic.*",
    "redis.*",
    "openai.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--strict-markers --tb=short -ra"
markers = [
    "asyncio: marks tests as async",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "smoke: marks tests as smoke tests",
    "slow: marks tests as slow",
    "critical: marks tests as critical path",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["core", "modulos", "canales", "onboarding", "seguridad", "billing", "config"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false
